black+isort; wildcard imports are now explicit

+++ cc/env/sample_from_spec.py
@@ -1,33 +1,36 @@
-from dm_env import specs as dm_env_specs 
-from beartype import beartype
 from collections import OrderedDict
-from beartype.typing import Union
-import jax 
-import numpy as np 
+
+import jax
+import numpy as np
 from acme.specs import Array
+from beartype import beartype
+from beartype.typing import Union
+from dm_env import specs as dm_env_specs
 
 
 def sample_action_from_action_spec(key, action_spec):
-    return jax.random.uniform(key, 
+    return jax.random.uniform(
+        key,
         shape=action_spec.shape,
         minval=action_spec.minimum,
         maxval=action_spec.maximum,
-        )
+    )
+
 
 ArraySpecs = Union[dm_env_specs.Array, dm_env_specs.BoundedArray]
+
 
 @beartype
 def sample_from_specs(specs: ArraySpecs):
     if isinstance(specs, dm_env_specs.Array):
-        return np.random.uniform(
-            low=-1e5, 
-            high=1e5,
-            size=specs.shape).astype(specs.dtype)
+        return np.random.uniform(low=-1e5, high=1e5, size=specs.shape).astype(
+            specs.dtype
+        )
     else:
         return np.random.uniform(
-            low=specs.minimum, 
-            high=specs.maximum,
-            size=specs.shape).astype(specs.dtype)
+            low=specs.minimum, high=specs.maximum, size=specs.shape
+        ).astype(specs.dtype)
+
 
 @beartype
 def sample_from_observation_specs(specs: Union[ArraySpecs, OrderedDict, dict]):
@@ -35,6 +38,7 @@ def sample_from_observation_specs(specs: Union[ArraySpecs, OrderedDict, dict]):
         return jax.tree_util.tree_map(lambda specs: sample_from_specs(specs), specs)
     else:
         return sample_from_specs(specs)
+
 
 def _spec_from_observation(observation):
     result = OrderedDict()
@@ -44,5 +48,3 @@ def _spec_from_observation(observation):
         else:
             result[key] = Array(value.shape, value.dtype, name=key)
     return result
-
-
